chore: Remove settings dependency on migrator
